Resolved issues in previous pull request with ConnectToWalletButton component.
