+++ package.json
@@ -22,7 +22,8 @@
     "react-scripts": "4.0.3",
     "react-window": "^1.8.6",
     "typescript": "^4.1.2",
-    "web-vitals": "^1.0.1"
+    "web-vitals": "^1.0.1",
+    "ethers": "^5.6.5"
   },
   "scripts": {
     "postinstall": "npm run apollo:download-schema && npm run apollo:codegen",
